Move MessageData type to separate module
