+++ webapp/src/utils/adapters/index.ts
@@ -1,0 +1,10 @@
+type TypeID = string;
+
+export type MessageData = {
+  defaultMessage: string;
+  id: string;
+  params: {
+    name: string;
+    types: TypeID[];
+  }[];
+};
